requirements.txt
asyncpg>=0.19.0
idna>=2.8
pylint>=2.4.3
pyOpenSSL>=19.0.0
PyYAML>=5.1.2
slackclient>=2.3.1
tornado>=6.0.3